Added -si, remove -ss
Firstly about -ss: I always thought that this function was useless and I
never used it. I just added it because when I made burgaur public I
thought that publishing a program with only two functions was bad. So I
removed it now that we have a search related function that is really
useful (this is -ss).

Now about -ss: this is a different implementation of PR#1. I really
liked the original idea but had some issues with the original
implementation, and the author seems to gave up on them. So I decided to
do it on my own.

This isn't the best implementation btw. I doesn't like how I am doing
the parsing of integers that the user entered, but it's 3AM and I am
tired, and this seems to work so whatever.