Replace python-blessings with python-termcolor
I quite liked python-blessings. They dynamic API is nice and beautiful to
use, but it has it's problems. For one, you need to create a global
object to use it. Not a problem until you need to create an optional
dependency on python-blessings... And I don't need its curses support
too.

python-termcolor is uglier in terms of API, but it does the job, it's
smaller and I can even create an optional dependency on it without too
much additional code. So now, color support is completelly optional (but
I still put it as depends on PKGBUILD because I think the majority of
people do want color support).