Update --color option parsing
The old code was not wrong, but the comment was old (still referring to python-blessings) and the checks was fragile (the user could pass garbage and it would assume "always" instead of the default "auto"). So now it displays a message to the user in case of garbage and assumes "auto", the default.